Fix IRawElementProviderSimple.GetPropertyValue (#9732)
Our interop handle structs (such as HWND) cannot be marshalled directly and will fail "silently" on
callbacks (they will throw but the marshaller will convert that to an HRESULT that we won't see unless first-chance exceptions are on when we're debugging).

+++ src/System.Windows.Forms/tests/UnitTests/System/Windows/Forms/ComboBox.ComboBoxAccessibleObjectTests.cs
@@ -272,7 +272,7 @@ public class ComboBox_ComboBoxAccessibleObjectTests
         object actual = comboBox.AccessibilityObject.GetPropertyValue(UiaCore.UIA.NativeWindowHandlePropertyId);
 
         Assert.True(comboBox.IsHandleCreated);
-        Assert.Equal(comboBox.InternalHandle, actual);
+        Assert.Equal((nint)comboBox.InternalHandle, actual);
     }
 
     [WinFormsTheory]
